move interface into actions file
https://coveord.atlassian.net/browse/KIT-661
