+++ packages/headless/src/controllers/facets/range-facet/numeric-facet/headless-numeric-facet.ts
@@ -3,12 +3,14 @@ import {
   NumericFacetRequest,
   NumericRangeRequest,
 } from '../../../../features/facets/range-facets/numeric-facet-set/interfaces/request';
-import {NumericFacetRegistrationOptions} from '../../../../features/facets/range-facets/numeric-facet-set/interfaces/options';
 import {
   NumericFacetResponse,
   NumericFacetValue,
 } from '../../../../features/facets/range-facets/numeric-facet-set/interfaces/response';
-import {registerNumericFacet} from '../../../../features/facets/range-facets/numeric-facet-set/numeric-facet-actions';
+import {
+  NumericFacetRegistrationOptions,
+  registerNumericFacet,
+} from '../../../../features/facets/range-facets/numeric-facet-set/numeric-facet-actions';
 import {
   assertRangeFacetOptions,
   buildRangeFacet,
