+++ packages/headless/src/features/facets/range-facets/generic/interfaces/range-facet.ts
@@ -6,7 +6,6 @@ import {
   NumericFacetRequest,
   NumericRangeRequest,
 } from '../../numeric-facet-set/interfaces/request';
-import {NumericFacetRegistrationOptions} from '../../numeric-facet-set/interfaces/options';
 import {
   DateFacetValue,
   DateFacetResponse,
@@ -16,6 +15,7 @@ import {
   NumericFacetResponse,
 } from '../../numeric-facet-set/interfaces/response';
 import {RegisterDateFacetActionCreatorPayload} from '../../date-facet-set/date-facet-actions';
+import {NumericFacetRegistrationOptions} from '../../numeric-facet-set/numeric-facet-actions';
 
 export type RangeFacetRequest = DateFacetRequest | NumericFacetRequest;
 export type RangeValueRequest = DateRangeRequest | NumericRangeRequest;
